// Codes/cagmss-web-a4/src/app/pages/models/newStation.ts
export class NewStation {
    cAera: string;
    cAgmesta: string;
    cAutosoilsta: string;
    cCity: string;
    cCode: string;
    cCropsta: string;
    cMeteosta: string;
    cPrvoName: string;
    cSoilsta: string;
    cStatName: string;
    cTypestation: string;
    v01000: number;
    v05001: number;
    v06001: number;
    v07001: number;
}