// Codes/cagmss-web-a4/src/app/pages/utils/models/mang_stat.ts
export class MangStat {
    v01000: number;
    cStatName: string;
    cPrvoName: string;
    v05001: number;
    v06001: number;
    v07001: string;
    cMeteoSta:string;
    cAgmeSta:string;
    cCropSta:string;
    cSoilSta:string;
    cAutoSoilSta:string;
    cTypestation: string;
    cAera: string;
    cCity: string;
    cCode: string;
}